move more build helpers to point of use
